backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
cssselect2==0.8.0
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
fastapi==0.121.2
filelock==3.20.0
fonttools==4.61.1
fpdf2==2.8.5
freetype-py==2.5.1
fsspec==2025.10.0
google-auth==2.43.0
google-genai==1.50.1
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
librt==0.7.8
limits==5.8.0
lxml==6.0.2
Markdown==3.10.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.8.0
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pathspec==1.0.4
pillow==12.1.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pymongo==4.15.4
pypdf==6.6.2
pytest==9.0.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
rpds-py==0.28.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.49.3
svglib==1.6.0
tabulate==0.9.0
tenacity==9.1.2
tinycss2==1.5.1
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
wrapt==2.1.1
xhtml2pdf==0.2.17
